+++ datafusion/core/src/datasource/mod.rs
@@ -21,7 +21,6 @@
 
 pub mod avro_to_arrow;
 pub mod cte_worktable;
-pub mod data_source;
 pub mod default_table_source;
 pub mod dynamic_file;
 pub mod empty;
