+++ datafusion/datasource/src/file.rs
@@ -15,18 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! DataSource and FileSource trait implementations
+//! Common behaviors that every file format needs to implement
 
 use std::any::Any;
 use std::fmt;
 use std::fmt::Formatter;
 use std::sync::Arc;
 
-use crate::datasource::physical_plan::{FileOpener, FileScanConfig};
-
+use crate::file_groups::FileGroupPartitioner;
+use crate::file_scan_config::FileScanConfig;
+use crate::file_stream::FileOpener;
 use arrow::datatypes::SchemaRef;
 use datafusion_common::Statistics;
-use datafusion_datasource::file_groups::FileGroupPartitioner;
 use datafusion_physical_expr::LexOrdering;
 use datafusion_physical_plan::metrics::ExecutionPlanMetricsSet;
 use datafusion_physical_plan::DisplayFormatType;
